bindOrder optimizations
Changed bindOrder values to be more consistent, both camelcased.
If you prefer you can use form_submit and sortable_stop, it's a matter of taste.

Added a select2SortableOrder() on original select change event in sortableStop binding, this is needed when you add the last option without sorting and than the first option without sorting. If you do not call select2SortableOrder() on change the values will be sent in the wrong order.

Removed a switch block, if bindOrder is not sortableStop it will fallback to form binding anyway.

I made this pull request on the fly editing the file in Github(I'm in hurry today) copying parts of my working code.
I didn't tested this mixed version but I'm confident it will works :), if you can't test it I will let you know if it works as soon as possible.